requirements.txt
Jinja2==3.1.4
Babel==2.14.0
pyyaml
markdown
